requirements.txt
orgo>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
requests>=2.25.0
threading-utils>=0.3 